requirements.txt
flask
sqlalchemy
mysql-connector-python
psycopg2-binary
flask-session
pymysql
bcrypt